refactor: pages styles, refactor on imports names
