+++ src/pages/auth/login/index.tsx
@@ -3,9 +3,9 @@ import Head from 'next/head';
 import React from 'react';
 import { QueryClient, QueryClientProvider } from 'react-query';
 import { AtLink, AtButton } from '../../../components/core/atoms';
-import { GuestLayout } from '../../../components/core/layouts';
+import { TmGuest } from '../../../components/core/templates';
 import { OrLoginForm } from '../../../components/core/organisms';
-import { LogoHeadSvg, GithubSvg } from '../../../components/utils/svg';
+import { LogoHeadIcon, GithubIcon } from '../../../components/utils/icons';
 import { RedirectionTypes, useNavigation } from '../../../hooks';
 import { handleGithubLogin } from '../handleGitHubLogin';
 import styles from './styles.module.scss';
@@ -25,16 +25,16 @@ const LoginPage: NextPage = () => {
         />
       </Head>
 
-      <GuestLayout>
+      <TmGuest>
         <div className={styles.loginContainer}>
           <div className={styles.logoContainer}>
             <div className={styles.logoBg}></div>
-            <LogoHeadSvg className={styles.logoHead} height={100} width={100} />
+            <LogoHeadIcon className={styles.logoHead} height={100} width={100} />
           </div>
 
           <div className={styles.loginForm}>
             <h2 className={styles.title}>Log in</h2>
-            <AtLink href={RedirectionTypes.REGISTER} label="Don't have an account?" classes="alignEnd" />
+            <AtLink href={RedirectionTypes.REGISTER} label="Don't have an account?" className="alignEnd" />
             <QueryClientProvider client={queryClient}>
               <OrLoginForm />
             </QueryClientProvider>
@@ -46,13 +46,13 @@ const LoginPage: NextPage = () => {
               <span>Or log in with</span>
               <hr />
             </div>
-            <AtButton onClick={handleGithubLogin(redirect)}>
-              <GithubSvg fill="#fff" />
+            <AtButton type="thirdParty" color="primary-light" onClick={handleGithubLogin(redirect)}>
+              <GithubIcon fill="#fff" />
               Github
             </AtButton>
           </div>
         </div>
-      </GuestLayout>
+      </TmGuest>
     </>
   );
 };
