+++ src/pages/auth/register/index.tsx
@@ -2,11 +2,11 @@ import styles from './styles.module.scss';
 import { NextPage } from 'next';
 import Head from 'next/head';
 import React from 'react';
-import { GithubSvg, LogoSvg, LogoHeadSvg } from '../../../components/utils/svg';
-import { GuestLayout } from '../../../components/core/layouts';
+import { GithubIcon, LogoIcon, LogoHeadIcon } from '../../../components/utils/icons';
+import { TmGuest } from '../../../components/core/templates';
 import { AtButton, AtLink } from '../../../components/core/atoms';
 import { OrRegisterForm } from '../../../components/core/organisms';
-import { RedirectionTypes, useNavigation } from 'hooks/useNavigation';
+import { RedirectionTypes, useNavigation } from '../../../hooks';
 import { QueryClient, QueryClientProvider } from 'react-query';
 import { handleGithubLogin } from '../handleGitHubLogin';
 
@@ -25,15 +25,15 @@ const RegisterPage: NextPage = () => {
         />
       </Head>
 
-      <GuestLayout>
+      <TmGuest>
         <div className={styles.logoContainer}>
           <div className={styles.logoBg}></div>
-          <LogoSvg className={styles.logo} height={320} width={320} />
-          <LogoHeadSvg className={styles.logoHead} height={100} width={100} />
+          <LogoIcon className={styles.logo} height={320} width={320} />
+          <LogoHeadIcon className={styles.logoHead} height={100} width={100} />
         </div>
         <div className={styles.registerContainer}>
           <h2 className={styles.title}>Sign up for free!</h2>
-          <AtLink href={RedirectionTypes.LOGIN} label="Already have an account?" classes="alignEnd" />
+          <AtLink href={RedirectionTypes.LOGIN} label="Already have an account?" className="alignEnd" />
           <QueryClientProvider client={queryClient}>
             <OrRegisterForm />
           </QueryClientProvider>
@@ -42,12 +42,12 @@ const RegisterPage: NextPage = () => {
             <span>Or sign up with</span>
             <hr />
           </div>
-          <AtButton onClick={handleGithubLogin(redirect)}>
-            <GithubSvg fill="#fff" />
+          <AtButton type="thirdParty" color="primary-light" onClick={handleGithubLogin(redirect)}>
+            <GithubIcon fill="#fff" />
             Github
           </AtButton>
         </div>
-      </GuestLayout>
+      </TmGuest>
     </>
   );
 };
