+++ src/pages/auth/register/styles.module.scss
@@ -65,7 +65,7 @@
   hr {
     height: 0;
     flex-grow: 1;
-    border: 1px solid #ccc;
+    border: 1px solid $gray;
   }
 
   span {
