+++ src/styles/pages/home.module.scss
@@ -5,12 +5,23 @@
 	-webkit-text-fill-color: transparent;
   margin: 0;
   line-height: 1.15;
-  font-size: 5rem;
-  text-align: center;
+  font-size: 3rem;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  flex-direction: column;
+  gap: 10px;
+
+  @media (min-width: $sm-width) {
+    flex-direction: row;
+    font-size: 5rem;
+  }
 }
 
 .slogan {
   font-weight: bold;
   font-size: 1.5rem;
   margin-bottom: 50px;
+  text-align: center;
+  width: 90%;
 }
